update like/unlike feedback

+++ Repository/IRepositories/ICourseRepository.cs
@@ -9,11 +9,11 @@ namespace Repository.IRepositories
 {
     public interface ICourseRepository : IBaseRepository<Course>
     {
-        Task<IEnumerable<Course>> GetAllSimilarCoursesAsync(long courseId, decimal score);
-        Task<Course?> GetByCourseIdAsync(long courseId);
-        Task<Course?> GetByCourseIdWithCategoryAsync(long courseId);
-        Task<IEnumerable<Course>> GetCoursesByCategoryAsync(long categoryId);
-        Task<(IEnumerable<Course> items, int total)> SearchAsync(int page, int pageSize, string? searchByName, long? couseId, CourseStatusEnum? status);
+        Task<IEnumerable<Course>> GetAllSimilarCoursesAsync(int courseId, decimal score);
+        Task<Course?> GetByCourseIdAsync(int courseId);
+        Task<Course?> GetByCourseIdWithCategoryAsync(int courseId);
+        Task<IEnumerable<Course>> GetCoursesByCategoryAsync(int categoryId);
+        Task<(IEnumerable<Course> items, int total)> SearchAsync(int page, int pageSize, string? searchByName, int? couseId, CourseStatusEnum? status);
         Task<Dictionary<string, int>> GetCoursesSortedByStatus();
         Task<IEnumerable<Course>> GetCoursesByCreatorAsync(string creatorAccountId);
         Task<CourseAnalyticsResponse> GetCourseAnalyticsAsync(int courseId);
